+++ Source/Modules/EJExtension/DelveEncounter/Bountiful/BountifulFrame.lua
@@ -57,8 +57,8 @@ function DelveCompanion_DelveEncounterBountifulFrameMixin:OnShow()
     self.ActiveDelves.NoBountifulLabel:SetShown(#self.ActiveDelves.Container:GetLayoutChildren() == 0)
 
     self.Title:ClearAllPoints()
-    if expansion == LE_EXPANSION_WAR_WITHIN then
-        self.DelveOBotWidget:SetShown(true) -- Delve-O-Bot 7001 works for TWW Delves only
+    if DelveCompanion.Variables.isPTR or expansion == LE_EXPANSION_WAR_WITHIN then
+        self.DelveOBotWidget:SetShown(true)
         self.Title:SetPoint("LEFT", self, "TOPLEFT", 5, -27)
         self.Title:SetJustifyH("LEFT")
     else